# openapi.yaml
openapi: 3.1.0
info:
  title: E-Business Card API
  version: 1.2.0
  description: Parses raw OCR text into hybrid-schema business card documents. Protected endpoints require a Firebase ID token.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
  /api/v1/cards:
    get:
      summary: List captured business cards for the authenticated user
      operationId: listCards
      tags:
        - cards
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cards owned by the authenticated user, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CapturedCardResponse"
        "401":
          description: Missing or invalid Firebase ID token
        "503":
          description: Authentication service not configured
        "500":
          description: Database query failed
  /api/v1/cards/process:
    post:
      summary: Parse OCR text and persist a captured business card
      operationId: processCard
      tags:
        - cards
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessCardRequest"
      responses:
        "201":
          description: Card parsed and saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapturedCardResponse"
        "401":
          description: Missing or invalid Firebase ID token
        "503":
          description: Authentication service not configured
        "502":
          description: LLM parsing service unavailable
        "504":
          description: LLM parsing service timed out
        "500":
          description: Database persistence failed
  /api/v1/user-cards:
    get:
      summary: List the authenticated user's own business cards
      operationId: listUserCards
      tags:
        - user-cards
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User cards, primary first then sort order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCardResponse"
        "401":
          description: Missing or invalid Firebase ID token
        "500":
          description: Database query failed
    post:
      summary: Create a user business card
      operationId: createUserCard
      tags:
        - user-cards
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCardCreate"
      responses:
        "201":
          description: User card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCardResponse"
        "401":
          description: Missing or invalid Firebase ID token
        "500":
          description: Database persistence failed
  /api/v1/user-cards/reorder:
    patch:
      summary: Reorder user business cards (first id becomes primary)
      operationId: reorderUserCards
      tags:
        - user-cards
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderUserCardsRequest"
      responses:
        "200":
          description: Updated user card list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCardResponse"
        "400":
          description: Invalid reorder payload
        "401":
          description: Missing or invalid Firebase ID token
  /api/v1/user-cards/parse:
    post:
      summary: Parse OCR text into user card fields (preview only)
      operationId: parseUserCard
      tags:
        - user-cards
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseUserCardRequest"
      responses:
        "200":
          description: Parsed fields preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedUserCardPreview"
        "401":
          description: Missing or invalid Firebase ID token
        "502":
          description: LLM parsing service unavailable
        "504":
          description: LLM parsing service timed out
  /api/v1/user-cards/{card_id}:
    put:
      summary: Update a user business card
      operationId: updateUserCard
      tags:
        - user-cards
      security:
        - BearerAuth: []
      parameters:
        - name: card_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCardUpdate"
      responses:
        "200":
          description: User card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCardResponse"
        "404":
          description: User card not found
        "401":
          description: Missing or invalid Firebase ID token
        "500":
          description: Database update failed
    delete:
      summary: Delete a user business card
      operationId: deleteUserCard
      tags:
        - user-cards
      security:
        - BearerAuth: []
      parameters:
        - name: card_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User card deleted
        "404":
          description: User card not found
        "401":
          description: Missing or invalid Firebase ID token
        "500":
          description: Database delete failed
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token from the mobile client
  schemas:
    ProcessCardRequest:
      type: object
      required:
        - raw_ocr_text
      properties:
        raw_ocr_text:
          type: string
          minLength: 1
    ParseUserCardRequest:
      type: object
      required:
        - raw_ocr_text
      properties:
        raw_ocr_text:
          type: string
          minLength: 1
    CoreFields:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        company_name:
          type: string
          nullable: true
        job_title:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
    ParsedUserCardPreview:
      type: object
      required:
        - core_fields
        - custom_fields
      properties:
        core_fields:
          $ref: "#/components/schemas/CoreFields"
        custom_fields:
          type: object
          additionalProperties:
            type: string
    CapturedCardResponse:
      type: object
      required:
        - _id
        - owner_user_id
        - scanned_at
        - core_fields
        - custom_fields
      properties:
        _id:
          type: string
        owner_user_id:
          type: string
          description: Firebase UID of the authenticated user
        scanned_at:
          type: string
          format: date-time
        core_fields:
          $ref: "#/components/schemas/CoreFields"
        custom_fields:
          type: object
          additionalProperties:
            type: string
    DesignType:
      type: string
      enum:
        - preset
        - custom
    UserCardCreate:
      type: object
      required:
        - core_fields
      properties:
        core_fields:
          $ref: "#/components/schemas/CoreFields"
        custom_fields:
          type: object
          additionalProperties:
            type: string
        design_id:
          type: string
          default: classic
        design_type:
          $ref: "#/components/schemas/DesignType"
        custom_background_url:
          type: string
          nullable: true
        is_primary:
          type: boolean
          default: false
    UserCardUpdate:
      type: object
      properties:
        core_fields:
          $ref: "#/components/schemas/CoreFields"
        custom_fields:
          type: object
          additionalProperties:
            type: string
        design_id:
          type: string
        design_type:
          $ref: "#/components/schemas/DesignType"
        custom_background_url:
          type: string
          nullable: true
        is_primary:
          type: boolean
    UserCardResponse:
      type: object
      required:
        - _id
        - owner_user_id
        - core_fields
        - custom_fields
        - design_id
        - design_type
        - is_primary
        - sort_order
        - created_at
        - updated_at
      properties:
        _id:
          type: string
        owner_user_id:
          type: string
        core_fields:
          $ref: "#/components/schemas/CoreFields"
        custom_fields:
          type: object
          additionalProperties:
            type: string
        design_id:
          type: string
        design_type:
          $ref: "#/components/schemas/DesignType"
        custom_background_url:
          type: string
          nullable: true
        is_primary:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReorderUserCardsRequest:
      type: object
      required:
        - ordered_ids
      properties:
        ordered_ids:
          type: array
          minItems: 1
          items:
            type: string